pyproject.toml
[tool.poetry]
name = "music-storage-api"
version = "0.1.0"
description = "Каталог исполнителей и их альбомов с песнями на основе Django Rest Framework"
authors = ["xander <moroz070688@gmail.com>"]
readme = "README.md"
packages = [{include = "music_storage_api"}]

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.4"
djangorestframework = "^3.14.0"
drf-yasg = "^1.21.7"
pre-commit = "^3.3.3"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
